pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["template-author"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "1.4.44"
Flask-SQLAlchemy = "3.0.2"
psycopg2 = "^2.9.7"
python-dotenv = "^1.0.0"
flask = {extras = ["async"], version = "^2.2.2"}
flask-restful = "^0.3.10"
marshmallow = "^3.20.1"
requests = "^2.31.0"
python-dateutil = "^2.8.2"
gunicorn = "^21.2.0"
celery = {extras = ["redis"], version = "^5.3.6"}
pulsar-client = {extras = ["avro"], version = "^3.4.0"}
pydispatcher = "^2.0.7"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
coverage = "^7.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
